Adds the matsuno-gill heating pattern as an idealized forcing namelist option. It is enabled in standalone HOMME by using the namelist option `test_case         = "matsuno_gill"`. This test case is described in, e.g., (this paper)[https://iopscience.iop.org/article/10.1016/j.fluiddyn.2004.03.003/pdf).
Add the vertical coordinate files for the Matsuno-Gill heating test.
The hybrid coefficients are analytically calculated to give constant ∆z between model levels for this test case,
with a model top at 10km.

Update CMakeLists.txt for theta-l_kokkos to include the new file for the matsuno gill heating profile.